Added lazy loading to some properties to fix a bug (in the description). Added comments on the news.
Fixed a bug with lazy loading on the sports of athletes because the sports weren't being displayed when loading the athletes page. They were displayed only when visiting a page that has the sports as a property. (For example the sports page or create a new athlete (because you set a sport there)).