Fixed validation for the update action, athlete controller (see description).
Fixed the validation on the update action in the athlete controller. The model state was giving false as a value to the IsValid property even though it should've given true when i removed the [Required} DataAnnotation in the UpdateAthleteViewModel. Now the sport id is nullable int, so the validation goes through and when you don't select a sport on update, the previous sport for the given athlete is being fetched from the database and inserted as data.